Update integration tests to run against AFT schema.
