+++ protomap/integration_tests/integration.go
@@ -1,0 +1,2 @@
+// Package integration_tests performs tests against real-world schemas for protomap.
+package integration_tests
