--- protomap/realtest/realtest.go
@@ -1,1 +1,0 @@
-package realtest
